Load thread's start routine from separate shared library
Prior to this change, if Lua unloads the threads library before the child
threads terminate, the program wil segfault. This could happen if an error
is triggered during an endcallback, preventing the child thread from exiting.

This places the thread's start routine in a separate shared library:
libthreadsmain, which is loaded once and never unloaded.